# src/main/java/com/sabrina/ecommerce/controller/api/OrderApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.ecommerce.controller.api;

import com.sabrina.ecommerce.controller.model.OrderApiResponse;
import com.sabrina.ecommerce.controller.model.OrderItemApiRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-01T11:52:41.153654499+02:00[Europe/Andorra]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Order", description = "the Order API")
public interface OrderApi {

    /**
     * POST /api/orders/{id}/items
     *
     * @param id  (required)
     * @param orderItemApiRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "addItemToOrder",
        tags = { "Order" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/orders/{id}/items",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<OrderApiResponse> addItemToOrder(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "OrderItemApiRequest", description = "", required = true) @Valid @RequestBody OrderItemApiRequest orderItemApiRequest
    );


    /**
     * POST /api/orders
     *
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "createOrder",
        tags = { "Order" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/orders",
        produces = { "application/json" }
    )
    
    ResponseEntity<OrderApiResponse> createOrder(
        
    );


    /**
     * GET /api/orders/my-orders
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getMyOrders",
        tags = { "Order" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OrderApiResponse.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/orders/my-orders",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<OrderApiResponse>> getMyOrders(
        
    );


    /**
     * GET /api/orders/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getOrderById",
        tags = { "Order" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/orders/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<OrderApiResponse> getOrderById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );

}